[KLIB IO] Extract the existing 'Path.deleteOnExitRecursively' utility function
^KT-87104
